[CIDMDEV-5129] - Migration Support - Read-Only Application (Service)
- Protected the following services by flag 'feature.migration.read.only.mode.enabled'
(returns 503 when read-only enabled) :
POST v2.0/OS-KSADM/services
DELETE v2.0/OS-KSADM/services/{serviceId}

+++ src/main/java/com/rackspace/idm/api/resource/cloud/v20/Cloud20VersionResource.java
@@ -741,7 +741,11 @@ public class Cloud20VersionResource {
             @Context HttpHeaders httpHeaders,
             @Context UriInfo uriInfo,
             @HeaderParam(X_AUTH_TOKEN) String authToken, Service service) {
-        return cloud20Service.addService(httpHeaders, uriInfo, authToken, service).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.addService(httpHeaders, uriInfo, authToken, service).build();
+        }
     }
 
     @GET
@@ -759,7 +763,11 @@ public class Cloud20VersionResource {
             @Context HttpHeaders httpHeaders,
             @HeaderParam(X_AUTH_TOKEN) String authToken,
             @PathParam("serviceId") String serviceId) {
-        return cloud20Service.deleteService(httpHeaders, authToken, serviceId).build();
+        if (identityConfig.getReloadableConfig().migrationReadOnlyEnabled()) {
+            return exceptionHandler.exceptionResponse(new MigrationReadOnlyIdmException()).build();
+        } else {
+            return cloud20Service.deleteService(httpHeaders, authToken, serviceId).build();
+        }
     }
 
     @GET
